feat: use abstract class for code-desc entities 😎
