+++ backend/src/main/java/ca/bc/gov/backendstartapi/entity/GameticMethodologyEntity.java
@@ -2,20 +2,13 @@ package ca.bc.gov.backendstartapi.entity;
 
 import ca.bc.gov.backendstartapi.entity.embeddable.EffectiveDateRange;
 import jakarta.persistence.Column;
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.beans.Transient;
-import java.time.Instant;
-import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 import lombok.Setter;
-import org.hibernate.annotations.SourceType;
-import org.hibernate.annotations.UpdateTimestamp;
 
 /** Entity for the list of gametic methodology */
 @Entity
@@ -23,14 +16,10 @@ import org.hibernate.annotations.UpdateTimestamp;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
 @Setter
-public class GameticMethodologyEntity {
+public class GameticMethodologyEntity extends CodeDescriptionEntity {
   @Id
   @Column(name = "gametic_methodology_code", length = 3)
   private String gameticMethodologyCode;
-
-  @Column(name = "description", length = 120)
-  @NonNull
-  private String description;
 
   @Column(name = "female_methodology_ind")
   private boolean isFemaleMethodology;
@@ -38,29 +27,15 @@ public class GameticMethodologyEntity {
   @Column(name = "pli_species_ind")
   private boolean isPliSpecies;
 
-  @Embedded @NonNull private EffectiveDateRange effectiveDateRange;
-
-  /** The date and time of the last update. */
-  @UpdateTimestamp(source = SourceType.DB)
-  private Instant updateTimestamp;
-
-  @Transient
-  public boolean isValid() {
-    var today = LocalDate.now();
-    return !effectiveDateRange.getEffectiveDate().isAfter(today)
-        && effectiveDateRange.getExpiryDate().isAfter(today);
-  }
-
   public GameticMethodologyEntity(
       String gameticMethodologyCode,
       String description,
       boolean isFemaleMethodology,
       boolean isPliSpecies,
       EffectiveDateRange effectiveDateRange) {
+    super(description, effectiveDateRange);
     this.gameticMethodologyCode = gameticMethodologyCode;
-    this.description = description;
     this.isFemaleMethodology = isFemaleMethodology;
     this.isPliSpecies = isPliSpecies;
-    this.effectiveDateRange = effectiveDateRange;
   }
 }
